+++ food_service/src/main/java/com/foodapi/food_service/repo/ProductRepo.java
@@ -3,6 +3,6 @@ package com.foodapi.food_service.repo;
 import com.foodapi.food_service.model.ProductModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepo extends JpaRepository<ProductModel,Integer > {
+public interface ProductRepo extends JpaRepository<ProductModel,Long > {
 
 }
